+++ aws/gen_mocks_test.go
@@ -32,6 +32,8 @@ import (
 	"github.com/aws/aws-sdk-go/service/cloudwatch/cloudwatchiface"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
+	"github.com/aws/aws-sdk-go/service/ecr"
+	"github.com/aws/aws-sdk-go/service/ecr/ecriface"
 	"github.com/aws/aws-sdk-go/service/elbv2"
 	"github.com/aws/aws-sdk-go/service/elbv2/elbv2iface"
 	"github.com/aws/aws-sdk-go/service/iam"
@@ -906,6 +908,62 @@ func (m *mockCloudformation) DescribeStacksPages(input *cloudformation.DescribeS
 	}
 	for i, page := range pages {
 		fn(&cloudformation.DescribeStacksOutput{Stacks: page, NextToken: aws.String(strconv.Itoa(i + 1))},
+			i < len(pages),
+		)
+	}
+	return nil
+}
+
+type mockEcr struct {
+	ecriface.ECRAPI
+	repositorys []*ecr.Repository
+}
+
+func (m *mockEcr) Name() string {
+	return ""
+}
+
+func (m *mockEcr) Provider() string {
+	return ""
+}
+
+func (m *mockEcr) ProviderAPI() string {
+	return ""
+}
+
+func (s *mockEcr) Drivers() []driver.Driver {
+	return []driver.Driver{
+		awsdriver.NewEcrDriver(s.ECRAPI),
+	}
+}
+
+func (m *mockEcr) ResourceTypes() []string {
+	return []string{}
+}
+
+func (m *mockEcr) FetchResources() (*graph.Graph, error) {
+	return nil, nil
+}
+
+func (m *mockEcr) IsSyncDisabled() bool {
+	return false
+}
+
+func (m *mockEcr) FetchByType(t string) (*graph.Graph, error) {
+	return nil, nil
+}
+
+func (m *mockEcr) DescribeRepositoriesPages(input *ecr.DescribeRepositoriesInput, fn func(p *ecr.DescribeRepositoriesOutput, lastPage bool) (shouldContinue bool)) error {
+	var pages [][]*ecr.Repository
+	for i := 0; i < len(m.repositorys); i += 2 {
+		page := []*ecr.Repository{m.repositorys[i]}
+		if i+1 < len(m.repositorys) {
+			page = append(page, m.repositorys[i+1])
+		}
+		pages = append(pages, page)
+	}
+	for i, page := range pages {
+		fn(&ecr.DescribeRepositoriesOutput{Repositories: page, NextToken: aws.String(strconv.Itoa(i + 1))},
 			i < len(pages),
 		)
 	}