+++ console/defaults.go
@@ -219,6 +219,14 @@ var DefaultsColumnDefinitions = map[string][]ColumnDefinition{
 		StringColumnDefinition{Prop: properties.AdjustmentType},
 		StringColumnDefinition{Prop: properties.ScalingAdjustment},
 	},
+	//Containers
+	cloud.Registry: {
+		StringColumnDefinition{Prop: properties.Name},
+		StringColumnDefinition{Prop: properties.URI},
+		TimeColumnDefinition{StringColumnDefinition: StringColumnDefinition{Prop: properties.Created}},
+		StringColumnDefinition{Prop: properties.Account},
+		StringColumnDefinition{Prop: properties.Arn},
+	},
 	//IAM
 	cloud.User: {
 		StringColumnDefinition{Prop: properties.ID},