+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## v0.1.1 [unreleased]
 
+## Features
+
 - `awless ls policies` now returns: your managed policies + all policies attached to any users, role or group
+- Listing [ECR](https://aws.amazon.com/ecr/) repositories: `awless list registries`
 
 ## v0.1.0 [2017-05-31]